# fleet_management_api/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: fleet@bringauto.com
    name: BringAuto s.r.o
    url: https://bringauto.com
  description: Specification for BringAuto fleet backend HTTP API
  license:
    name: AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  title: BringAuto Fleet Management v2 API
  version: 3.0.0
servers:
- url: /v2/management
security:
- APIKeyAuth: []
- oAuth2AuthCode: []
...
paths:
...
  /order:
    get:
      operationId: get_orders
      parameters:
      - description: "A Unix timestamp in milliseconds. If specified, only objects\
          \ created at the time or later will be returned. If unspecified, all objects\
          \ are returned (since is set to 0 in that case)."
        in: query
        name: since
        schema:
          format: int64
          type: integer
      responses:
        "200":
          content:
            application/json:
              example:
              - priority: normal
                userId: 1
                carId: 1
                timestamp: 1713256508978
                targetStopId: 1
                notification: Order notification
                stopRouteId: 1
                notificationPhone:
                  phone: "+420123456789"
              schema:
                items:
                  $ref: '#/components/schemas/Order'
                type: array
          description: All the currently existing Orders have been sorted by their
            creation timestamp from the oldest to newest and returned.
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Find all currently existing Orders.
      tags:
      - order
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.order
    post:
      operationId: create_orders
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/Order'
              type: array
        description: Order model in JSON format.
        required: true
        x-body-name: order
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Order'
                type: array
          description: The new Order has been successfully created.
        "400":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "400"
                    message: Bad request
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Create a new Order.
      tags:
      - order
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.order
  /order/{carId}:
    get:
      operationId: get_car_orders
      parameters:
      - description: "A Unix timestamp in milliseconds. If specified, only objects\
          \ created at the time or later will be returned. If unspecified, all objects\
          \ are returned (since is set to 0 in that case)."
        in: query
        name: since
        schema:
          format: int64
          type: integer
      - description: ID of the Car for which Orders shall be returned.
        example: 1
        in: path
        name: carId
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Order'
                type: array
          description: "The Orders assigned to the Car with the given ID have been\
            \ found, sorted by their creation timestamp from oldest to newest and\
            \ returned."
        "400":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "400"
                    message: Bad request
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "404"
                    message: Resource not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Find existing Orders by the corresponding Car ID and return them.
      tags:
      - order
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.order
  /order/{carId}/{orderId}:
    delete:
      operationId: delete_order
      parameters:
      - description: ID of the Car to which the Order is assigned.
        example: 1
        in: path
        name: carId
        required: true
        schema:
          format: int32
          type: integer
      - description: ID of the Order to be returned.
        example: 1
        in: path
        name: orderId
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          content:
            text/plain: {}
          description: The Order with the specified ID has been successfully deleted.
        "400":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "400"
                    message: Bad request
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "404"
                    message: Resource not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
        "405":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "405"
                    message: Method not allowed
              schema:
                $ref: '#/components/schemas/Error'
          description: Method not allowed
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Delete an Order identified by its ID and ID of a car to which it is
        assigned.
      tags:
      - order
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.order
    get:
      operationId: get_order
      parameters:
      - description: ID of the Car to which the Order is assigned.
        example: 1
        in: path
        name: carId
        required: true
        schema:
          format: int32
          type: integer
      - description: ID of the Order to be returned.
        example: 1
        in: path
        name: orderId
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
          description: The Order with the specified car ID and order ID has been found
            and returned.
        "400":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "400"
                    message: Bad request
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "404"
                    message: Resource not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Find an existing Order by the car ID and the order ID and return it.
      tags:
      - order
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.order
...
components:
...
  schemas:
    Error:
      description: Error object structure.
      properties:
        code:
          example: 404
          title: code
          type: integer
        message:
          example: Not found
          title: message
          type: string
      required:
      - code
      - message
      title: Error
      type: object
...
    Order:
      description: Order object structure.
      example:
        notification: Order notification
        stopRouteId: 1
        notificationPhone:
          phone: "+420123456789"
        targetStopId: 1
        id: 1
        priority: normal
        lastState:
          orderId: 1
          id: 1
          status: to_accept
          timestamp: 1616425275913
        userId: 1
        timestamp: 1616425275913
        carId: 1
      properties:
        id:
          example: 1
          format: int32
          title: Id
          type: integer
        priority:
          default: normal
          description: "Priority (low, normal, high)"
          pattern: ^(low|normal|high)$
          title: priority
          type: string
        userId:
          example: 1
          format: int32
          title: Id
          type: integer
        timestamp:
          description: A Unix timestamp in milliseconds. The timestamp is used to
            determine the time of creation of an object.
          example: 1616425275913
          format: int64
          title: Timestamp
          type: integer
        carId:
          example: 1
          format: int32
          title: Id
          type: integer
        notification:
          example: Order notification
          title: notification
          type: string
        targetStopId:
          example: 1
          format: int32
          title: Id
          type: integer
        stopRouteId:
          example: 1
          format: int32
          title: Id
          type: integer
        notificationPhone:
          $ref: '#/components/schemas/MobilePhone'
        lastState:
          $ref: '#/components/schemas/OrderState'
      required:
      - carId
      - stopRouteId
      - targetStopId
      - userId
      title: Order
      type: object
...
    OrderState:
      description: Order state object structure.
      example:
        orderId: 1
        id: 1
        status: to_accept
        timestamp: 1616425275913
      properties:
        id:
          example: 1
          format: int32
          title: Id
          type: integer
        status:
          $ref: '#/components/schemas/OrderStatus'
        orderId:
          example: 1
          format: int32
          title: Id
          type: integer
        timestamp:
          description: A Unix timestamp in milliseconds. The timestamp is used to
            determine the time of creation of an object.
          example: 1616425275913
          format: int64
          title: Timestamp
          type: integer
      required:
      - orderId
      - status
      title: OrderState
      type: object
    OrderStatus:
      description: "OrderStatus (to_accept, accepted, in_progress, done, canceled)."
      enum:
      - to_accept
      - accepted
      - in_progress
      - done
      - canceled
      example: to_accept
      title: OrderStatus
      type: string
...
    MobilePhone:
      description: MobilePhone Primitive structure.
      example:
        phone: "+420123456789"
      properties:
        phone:
          example: "+420123456789"
          title: phone
          type: string
      title: MobilePhone
      type: object
...
  securitySchemes:
    APIKeyAuth:
      description: Authentication using an API key.
      in: query
      name: api_key
      type: apiKey
      x-apikeyInfoFunc: fleet_management_api.controllers.security_controller.info_from_APIKeyAuth
    oAuth2AuthCode:
      description: oAuth2 authorization
      flows:
        authorizationCode:
          authorizationUrl: https://keycloak.bringauto.com/realms/bringauto/protocol/openid-connect/auth
          refreshUrl: https://keycloak.bringauto.com/realms/bringauto/protocol/openid-connect/token
          scopes: {}
          tokenUrl: https://keycloak.bringauto.com/realms/bringauto/protocol/openid-connect/token
      type: oauth2
      x-tokenInfoFunc: fleet_management_api.controllers.security_controller.info_from_oAuth2AuthCode
      x-scopeValidateFunc: fleet_management_api.controllers.security_controller.validate_scope_oAuth2AuthCode